created commands and api modules and refactored code
