+++ go.mod
@@ -3,8 +3,11 @@ module github.com/jchu47/OpenCLIDocs
 go 1.21.4
 
 require (
+	github.com/sashabaranov/go-openai v1.17.9
+	github.com/spf13/cobra v1.8.0
+)
+
+require (
 	github.com/inconshreveable/mousetrap v1.1.0 // indirect
-	github.com/sashabaranov/go-openai v1.17.9 // indirect
-	github.com/spf13/cobra v1.8.0 // indirect
 	github.com/spf13/pflag v1.0.5 // indirect
 )
